requirements.txt
alembic==1.8.1
bcrypt==4.0.0
certifi==2022.9.24
charset-normalizer==2.1.1
click==8.1.3
Flask==2.2.2
Flask-Bcrypt==1.0.1
Flask-JWT-Extended==4.4.4
Flask-Migrate==3.1.0
Flask-SQLAlchemy==2.5.1
greenlet==1.1.3
idna==3.4
importlib-metadata==4.12.0
importlib-resources==5.9.0
itsdangerous==2.1.2
Jinja2==3.1.2
Mako==1.2.3
MarkupSafe==2.1.1
psycopg2-binary==2.9.3
psycopg2==2.9.3
PyJWT==2.5.0
python-decouple==3.6
python-dotenv==0.21.0
requests==2.28.1
SQLAlchemy==1.4.41
urllib3==1.26.12
Werkzeug==2.2.2
zipp==3.8.1
